requirements.txt
applicationinsights>=0.11.7
colour>=0.1.5
chardet>=3.0.4
coverage>=4.5.3
dj-database-url>=0.5.0
Django>=2.1.11
django-cloud-browser>=0.5.0
django-cors-headers>=3.1.1
django-environ>=0.4.5
django-filter>=2.0.0
django-webpack-loader>=0.6.0
django-widget-tweaks>=1.4.2
django-polymorphic>=2.0.3
# django-pyodbc-azure>=2.1.0.0
django-rest-polymorphic>=0.1.8
djangorestframework>=3.10
djangorestframework-csv>=2.1.0
djangorestframework-filters>=0.10.2
environs>=4.1.0
djangorestframework-xml>=1.4.0
Faker>=0.9.1
flake8>=3.6.0
furl>=2.0.0
gunicorn>=19.9.0
lockfile>=0.12.2
model-mommy>=1.6.0
psycopg2-binary>=2.7.7
pyexcel>=0.5.14
pyexcel-xlsx>=0.5.7
python-dateutil>=2.7.3
pytz>=2018.4
requests>=2.21.0
six>=1.11.0
seqeval>=0.0.6
text-unidecode>=1.2
unittest-xml-reporting>=2.5.1
vcrpy>=2.0.1
vcrpy-unittest>=0.1.7
whitenoise[brotli]>=4.1.2
conllu>=1.3.2
spacy
pyzmq >= 17.1.0
termcolor
flair>=0.4.3
pandas
